requirements.txt
aiohttp==3.7.4.post0
anyio==3.2.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
discord-pretty-help==1.3.3
discord.py==1.7.3
googlesearch-python==1.0.1
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==2.10
imgurpython==1.1.7
joblib==1.0.1
multidict==5.1.0
mysql-connector-python==8.0.28
nltk==3.6.2
Pillow==9.0.0
praw==7.3.0
prawcore==2.2.0
protobuf==3.19.3
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.18.0
regex==2021.7.6
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
textblob==0.15.3
tqdm==4.61.2
typing-extensions==3.10.0.0
update-checker==0.18.0
urllib3==1.26.6
websocket-client==1.1.0
wikipedia==1.4.0
xkcd==2.4.2
yarl==1.6.3
youtube-dl==2021.6.6
youtube-search-python==1.4.6
